# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/skill-template/generate": {
      "post": {
        "description": "根据名称和场景生成可复用的 skill 草稿模板。",
        "operationId": "SkillTemplateController_generate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSkillTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功（统一响应体：code/status/message/data/ts）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillTemplateResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "生成 Skill 模板",
        "tags": [
          "Skill 模板模块"
        ]
      }
    },
    "/skill-template/list": {
      "get": {
        "description": "按关键字与状态过滤模板列表。",
        "operationId": "SkillTemplateController_list",
        "parameters": [
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键字过滤",
            "schema": {
              "example": "解析源",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态过滤",
            "schema": {
              "example": "draft",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功（统一响应体：code/status/message/data/ts）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SkillTemplateResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Skill 模板列表",
        "tags": [
          "Skill 模板模块"
        ]
      }
    },
    "/skill-template/detail": {
      "post": {
        "description": "单对象查询统一使用 body.id。",
        "operationId": "SkillTemplateController_detail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillTemplateIdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "查询成功（统一响应体：code/status/message/data/ts）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillTemplateResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Skill 模板详情",
        "tags": [
          "Skill 模板模块"
        ]
      }
    }
  },
  "info": {
    "title": "Isle Flix 服务端接口文档",
    "description": "视频资源聚合服务接口文档（中文）。单对象操作统一使用 id，禁止路径参数。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "先调用 /auth/login 获取 accessToken，再在此处填入。"
      }
    },
    "schemas": {
...
      "GenerateSkillTemplateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "skill 模板名称",
            "example": "解析源切换规范"
          },
          "scenario": {
            "type": "string",
            "description": "适用场景描述",
            "example": "用户需要在客户端切换解析源时使用。"
          }
        },
        "required": [
          "name",
          "scenario"
        ]
      },
      "SkillTemplateResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "模板主键 id",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "description": "模板名称",
            "example": "解析源切换规范"
          },
          "scenario": {
            "type": "string",
            "description": "场景说明",
            "example": "用户需要在客户端切换解析源时使用。"
          },
          "status": {
            "type": "string",
            "description": "模板状态",
            "example": "draft"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2026-05-14T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "scenario",
          "status",
          "createdAt"
        ]
      },
      "SkillTemplateIdDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "skill 模板主键 id（单 id 场景统一使用 id）",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        },
        "required": [
          "id"
        ]
      }
    }
  }
}